fix replyCntroller wrong place problem
